+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added:
+- Support for inline code in editor
+- Support for strikethrough inline style in editor
+
+## [0.8.85] - 2022-06-30
 ### Changed:
 - FadeIn animation on new measurement page